Remove Registrable feature (it seems not to be a good idea)
Studying functional programming, I learnt that the mathematical theories do not
define any mutable operations. So, no programming language or pattern (or
framework) should provide it as well.

In this way, value objects should not be muttable and Bauhaus framework must not
provide a mutable container, once it will be the superclass of many VO.

Another argument is that the interface to register a new item is responsability
of the readable container subclass.

References:
- https://www.coursera.org/learn/progfun1
- http://martinfowler.com/bliki/ValueObject.html